pyproject.toml
[project]
name = "vigilens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiohttp",
  "boto3",
  "botocore",
  "fastapi",
    "google-genai",
  "httpx",
  "modal>=1.3.3",
  "numpy",
  "opencv-python-headless",
  "pillow",
  "pydantic",
  "pydantic-settings",
  "pyyaml",
  "qwen-vl-utils",
  "redis",
  "tenacity>=9.1.4",
  "torch",
  "transformers",
  "uvicorn",
]

[project.optional-dependencies]
macdev = []
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "aioresponses>=0.7",
    "fakeredis[aioredis]>=2.21",
    "httpx>=0.27",
    "aiosqlite>=0.20",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "asyncio: mark a test as an asyncio coroutine",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "pre-commit>=4.5.1",
    "ruff>=0.15.2",
]
